Moving the multi line fixes out to a seperate rule so that we can ensure we aren't doing strange stuff cloning nodes, etc
