+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/AnnotationRule.kt
@@ -9,9 +9,7 @@ import com.pinterest.ktlint.core.ast.ElementType.VALUE_PARAMETER
 import com.pinterest.ktlint.core.ast.isPartOf
 import com.pinterest.ktlint.core.ast.isPartOfComment
 import com.pinterest.ktlint.core.ast.isWhiteSpace
-import com.pinterest.ktlint.core.ast.isWhiteSpaceWithNewline
 import com.pinterest.ktlint.core.ast.lineNumber
-import com.pinterest.ktlint.core.ast.nextLeaf
 import com.pinterest.ktlint.core.ast.nextSibling
 import com.pinterest.ktlint.core.ast.prevSibling
 import com.pinterest.ktlint.core.ast.upsertWhitespaceBeforeMe
@@ -31,6 +29,8 @@ import org.jetbrains.kotlin.psi.psiUtil.nextLeaf
  * with parameters are placed on separate lines.
  *
  * https://kotlinlang.org/docs/reference/coding-conventions.html#annotation-formatting
+ *
+ * @see [MultiLineAnnotationRule] for white space rules. Moved since
  */
 class AnnotationRule : Rule("annotation") {
 
@@ -41,8 +41,6 @@ class AnnotationRule : Rule("annotation") {
             "Annotations with parameters should all be placed on separate lines prior to the annotated construct"
         const val fileAnnotationsShouldBeSeparated =
             "File annotations should be separated from file contents with a blank line"
-        const val fileAnnotationsLineBreaks =
-            "There should not be empty lines between an annotation and the object that it's annotating"
     }
 
     override fun visit(
@@ -144,67 +142,6 @@ class AnnotationRule : Rule("annotation") {
                 }
             }
         }
-
-        // Check to make sure no trailing line breaks between annotation and object
-        val lineNumber = node.lineNumber()
-        val next = node.nextSiblingWithAtLeastOneOf(
-            {
-                !it.isWhiteSpace() &&
-                    it.textLength > 0 &&
-                    !(it.isPartOfComment() && it.lineNumber() == lineNumber) &&
-                    !it.isPartOf(FILE_ANNOTATION_LIST)
-            },
-            {
-                val s = it.text
-                // Ensure at least one occurrence of two line breaks
-                s.indexOf("\n") != s.lastIndexOf("\n")
-            }
-        )
-        val nextLineNumber = next?.lineNumber()
-        if (lineNumber != null && nextLineNumber != null) {
-            val diff = nextLineNumber - lineNumber
-            // Ensure declaration is not on the same line, there is a line break in between, and it is not an
-            // annotation we explicitly want to have a line break between
-            if (diff > 1 && node.elementType != FILE_ANNOTATION_LIST) {
-                val psi = node.psi
-                emit(psi.endOffset - 1, fileAnnotationsLineBreaks, true)
-                if (autoCorrect) {
-                    removeExtraLineBreaks(node)
-                }
-            }
-        }
-        if (whiteSpaces.isNotEmpty() && annotations.size > 1 && node.elementType != FILE_ANNOTATION_LIST) {
-            // Check to make sure there are multi breaks between annotations
-            if (whiteSpaces.any { psi -> psi.textToCharArray().filter { it == '\n' }.count() > 1 }) {
-                val psi = node.psi
-                emit(psi.endOffset - 1, fileAnnotationsLineBreaks, true)
-                if (autoCorrect) {
-                    removeIntraLineBreaks(node, annotations.last())
-                }
-            }
-        }
-    }
-
-    private inline fun ASTNode.nextSiblingWithAtLeastOneOf(
-        p: (ASTNode) -> Boolean,
-        needsToOccur: (ASTNode) -> Boolean
-    ): ASTNode? {
-        var n = this.treeNext
-        var occurrenceCount = 0
-        while (n != null) {
-            if (needsToOccur(n)) {
-                occurrenceCount++
-            }
-            if (p(n)) {
-                return if (occurrenceCount > 0) {
-                    n
-                } else {
-                    null
-                }
-            }
-            n = n.treeNext
-        }
-        return null
     }
 
     private fun getNewlineWithIndent(modifierListRoot: ASTNode): String {
@@ -220,45 +157,8 @@ class AnnotationRule : Rule("annotation") {
         }
     }
 
-    private fun removeExtraLineBreaks(node: ASTNode) {
-        val next = node.nextSibling {
-            it.isWhiteSpaceWithNewline()
-        } as? LeafPsiElement
-        if (next != null) {
-            rawReplaceExtraLineBreaks(next)
-        }
-    }
-
-    private fun rawReplaceExtraLineBreaks(leaf: LeafPsiElement) {
-        // Replace the extra white space with a single break
-        val text = leaf.text
-        val firstIndex = text.indexOf("\n") + 1
-        val replacementText = text.substring(0, firstIndex) +
-            text.substringAfter("\n").replace("\n", "")
-
-        leaf.rawReplaceWithText(replacementText)
-    }
-
     private fun doesNotEndWithAComment(whiteSpaces: List<PsiWhiteSpace>): Boolean {
         val lastNode = whiteSpaces.lastOrNull()?.nextLeaf()
         return lastNode !is PsiComment || lastNode.nextLeaf()?.textContains('\n') == false
-    }
-
-    private fun removeIntraLineBreaks(
-        node: ASTNode,
-        last: KtAnnotationEntry
-    ) {
-        val txt = node.text
-        // Pull the next before raw replace or it will blow up
-        val lNext = node.nextLeaf()
-        if (node is PsiWhiteSpaceImpl) {
-            if (txt.toCharArray().count { it == '\n' } > 1) {
-                rawReplaceExtraLineBreaks(node)
-            }
-        }
-
-        if (lNext != null && !last.text.endsWith(lNext.text)) {
-            removeIntraLineBreaks(lNext, last)
-        }
     }
 }
